content/base/test/test_bug600466.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=600466
-->
<head>
  <title>Test for Bug 600466</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=600466">Mozilla Bug 600466</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 600466 **/

var attr = document.createAttribute("foo");
attr.value = "bar";
ok(attr.firstChild, "attr should have a first child");

attr.removeChild(attr.firstChild);
ok(!attr.firstChild, "attr should not have a first child anymore");

</script>
</pre>
</body>
</html>
